Version with some new features:
a) One should put other two arguments in command line: the number of
snapshots and images outputted during the pulsatile cycle (actually
the images are not outputted because of a bug due to the network/vis
thread part.

b) streaklines: particles periodically generated by each inlet. The
colour of the particle depends on the inlet id is generated from and
the current particle velocity. The rendering is merged with the
visualisation of the stress and pressure at the wall; the visualisation is active if "vis_glyph_length" is greater or equal 0, otherwise glyphs are merged with those vis modalities.

c) for each inlet, the calculation of its flow rate times its area
(eventually calculated with some software) is performend during the
pulsatile cycle; if the normals of the inlet boundaries are written at
the end of the file pars.asc (available with the new version of the
segmentation tool seg_tool.cc) the flux computation is more accurate.

d) the snapshots are outputted in a more compact way if the minimum
bounding box of the system is smaller than the bounding box of the
current data set. This saves memory during post-processing.