Hi,
So we worked out what the issue was.

On MAC OS X the call to MouseEvent.e.getButton() returns 1,2,3 depending on which button you press. This is the same on other linux/java(?) version.

However, on RH Enterprise linux 4 with the latest Sun Java SDK it always returns 0 irrespective of the mouse button pressed.

And Stefan found this -

according to the JavaDoc "Due to platform-dependent Drag&Drop implementations, MOUSE_DRAGGED events may not be delivered during a native Drag&Drop operation."

Java, the platform independent language.

Fun stuff.

Steven.