CfgParser: use weakrefs to ensure memory freed asap
Use the weakref standard library to ensure that the subdomains created for the worker threads to check are deleted as soon as possible (without having to wait for a GC cycle detector pass to run). Since multiprocessing serialises objects with pickle, and weakrefs are not pickable, Site and Block have to implement custom __getstate__ and __setstate__ methods that translate the weakref to a strong ref (get) and back to weakrefs (set)

Also change the API for the worker function, now called with a (Domain, blockIndex) instead of just a Block.