Factor util/Vector3D.h so it's ready for use by the setuptool (#115)
The setuptool doesn't use MPI, so here I put the necessary template
specialisations in util/Vector3DHemeLb.cc. Factor the index error
handling into a base class, with a static method as the handler. This
delegates to a class static member variable (function pointer) that is
initialised (in HemeLB) in util/Vector3DHemeLb.cc

The setuptool's Vec3 had some methods that were not implemented in
Vector3D, so implement them.

Give proper Doxygen comments to everything.

Rename a few methods and make the appropriate updates in HemeLB.