Rewrite the setuptool geometry writing (#98)
This basically works now and is a great deal clearer.

Remove redundant member var BlockWriter::usedBufferSize

Cache the vtkIntArray that stores the mapping from triangles of the
surface to iolet IDs in ConfigGenerator::IoletIdArray

Factor the writing of different types of site into private methods of ConfigGenerator

Make ConfigGenerator::ClassifySite private

The Neighbours class is simpler now. The later-ness and inverses are computed during Neighbours::Init

The Site class is much simplified, but a simple struct LinkData has been added.