SetupTool: make the python layer cope with change to working in voxels (#98)
This alters the use of the wrapped C++ by the python layer to cope
with its change in naming. It also removes the scaling done before
'Generate' is triggered to simplify the code. The input data is now
scaled at the last moment before handing over to C++, along with all
the positions of inlets/outlets etc. The labelling of fields in the UI
has be updated to just mirror the selected units of the file, instead
of always being metres.

--HG--
branch : newGeometryFormat