#104: add uncompressed block data length to geometry file format
To do this, I have added the necessary to the setuptool, hemeTools
parser, four_cube generator and the main application. In the main, I
have updated formats/geometry.h (the HeaderRecordLength) and the
GeometryReader class. In the latter, Miguel and I refactored
sitesPerBlock and bytesPerBlock into member variables (renaming to
fluidSitesPerBlock and bytesPerCompressedBlock respectively) and added
bytesPerUncompressedBlock. We did not move all of the usages to these
member vars from function args.