Refactor the streamers using delegates to aid code reuse
The design is to for the streamers to be pretty dumb and for them to
basically just control iteration over the sites and directions (to be
pulled into a common place in a future refactor) but to delegate the
logic of actually streaming to some other classes
(e.g. BouzidiFirdaousLallemand delegates bulk link streaming to
SimpleCollideAndStreamDelegate and wall link streaming to BFLDelagate,
which uses SimpleBounceBackDelegate in the cases where it can't handle
because two opposite links are both wall links). Since the methods
that do the work are NOT polymorphic (and marked as inline) the
compiler can inline them quite straightforwardly---I checked the
assembly output on gcc with -O2 and this is in fact done. The overhead
should be pretty low.