Pull common streamer code into templates, use metafunctions to do partial template specialisation.
Much of the streamers code is basically identical between classes but
is repeated, particularly the loop over sites, calling the collider,
looping over directions doing the modified stream and then updating
the macroscopic variables. Factor this into a small number
of templates (in StreamerTypeFactor.h) which are then partially
specialised on the different boundary conditions in simple
metafunctions (in e.g. LaddIolet.h). Update the build system to
reflect this.