Adding bending energy implementation
Currently only between neighboring facets: no loops over all facets yet.
None of the implementation is actually exported. It remains within an anonymous
namespace. For unit-testing purposes, the anonymous namespaces is contained in
a separate file that is included in the tests, and will be included in the main
Particle soon to be declared.