Tested velocity interpolation
Velocity from LB grid are interpolated onto the nodes of the membrane.
The main usage pattern should be something of the form:

~~~cpp
LatticeVelocity vel = interpolateVelocity(velocityFunctor, node_position, stencil);
~~~

If missing, ``stencil`` defaults to the four point stencil. The velocityFunctor
is an object which given a 3d global index into the LB grid returns the
velocity (or Guo corrected velocity) at that point. An typical implementation
is given in ``Code/redblood/VelocityInterpolation.h``.

Once we have merged with the colloid branch, a functor will be created that
computes the Guo (force-corrected) velocity. A functor could also be created to
use the cached momentum.