Jenkins jobs reunited
Dependencies, build, unit-testing, regression all under same job.
As well as compilation for different flags.

The reason is that we cannot be sure that dependent jobs are on the same slave.
So build job might not find stuff it needs from dependencies build. And
regression won't find what it needs from build job...