Buffer drops cells upon request if possible
Usage is to first make a request via Buffer::requestNewCell(site_t), then to actually ask the buffer
to drop those cells if there is enough space available in the drop zone, via
Buffer::operator()(callback). The callback takes a new cell as input and does what it does with new
cells.

This commit should close