Adding function to orient mesh facets
Orientation is done by creating a scaled copy (smaller in this case) of the mesh with the same
center of mass. We can determine the orientation of a facet by comparing its normal to the vector
between the center of the original facet and the center of the eqivalent scaled facet.

It turns out that Timm's mesh was not oriented outward.
This is #702