Incorrect lambda + static variable
I wanted a variable that would live from one call of the lambda to the next, e.g. a member variable
the lambda instance. I was using a static variable within the lambda call. This is actually
equivalent to a static variable of the lambda class. So it is shared by all instances of the lambda
class. The lambda class is defined for each lambda within the scope where the lambda is defined. So
once per factory function.

C++14 will allow more sophisticated captures (e.g. with initialization). Until then, we will have to
use a shared pointer to emulate a variable of the lambda instance.