Introduce a map that allows different data sources for property output.
NOTE: At the moment, it is not possible to have different data sources for the same output file. This requires some more changes in the Write() methods. In principle, each hemelb::extraction::OutputField should have its own dataSource and OutputField should be virtual class. Each specific output can then be derived from OutputField and implement the proper Write() methods. This would eliminate the need for all that switch-case shenanigans.

This is a not so elegant but unavoidable workaround for the lack of abstraction in the PropertyWriter. Essentially, we introduce a std::map<OutputField::FieldType, IterableDataSource*>& that is passed through to hemelb::extraction::LocalPropertyOutput where the constructor initialises the dataSource according to the map. In principle, this information should be encapsulated in hemelb::extraction::PropertyOutputFile (which should really be called PropertyOutputDescriptor), however, during initialisation of PropertyOutputFile the propertyCaches are yet uniinitalised. Ultimately, the whole simulation master should be re-designed to have a proper distinction between allocation and initialisation.